+++ Task09/app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Task09</string>
+    <string name="img_description1">image 1</string>
 </resources>
