+++ Task10/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 27
     defaultConfig {
-        applicationId "com.example.will.task09"
+        applicationId "com.example.will.task10"
         minSdkVersion 14
         targetSdkVersion 27
         versionCode 1
@@ -15,16 +15,6 @@ android {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
-    }
-    lintOptions {
-        // lintファイルを直指定するため
-        lintConfig rootProject.file("lint.xml")
-        // Errorを出力するため
-        abortOnError true
-        // エラー箇所を出力するため
-        noLines false
-        // エラー内容を出力するため
-        textReport true
     }
 }
