+++ Task10/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Thu Apr 05 15:09:26 JST 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.4-all.zip
